requirements.txt
pandas==1.4.1
slackclient==2.9.3
requests==2.23.0